% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-package.R
\docType{package}
\name{fmp-package}
\alias{fmp}
\alias{fmp-package}
\title{fmp: Access to the Financial Modeling Prep Financial Data API}
\description{
simple and consistent interface to the Financial Modeling Prep Financial Data API
}
\keyword{internal}
