% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutional-funds.R
\name{fmp_13f}
\alias{fmp_13f}
\title{Form 13-F Statements}
\usage{
fmp_13f(cik, year = NULL)
}
\arguments{
\item{cik}{\code{character}. A vector of Central Index Keys (CIK)}

\item{year}{\code{numeric}. 4 digit year, e.g. \code{2020}. If \code{NULL}, the current year
will be taken from \code{\link[=Sys.Date]{Sys.Date()}}. Default is \code{NULL}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Form 13-F statements provides position-level report of all institutional
investment managers with more than $100m in assets under management
}
\examples{

berkshire_cik <- '0001067983'
\donttest{
d <- fmp_13f(berkshire_cik)
}

\donttest{
# 13-F forms from 2018
d_18 <- fmp_13f(berkshire_cik, year = 2018)
}

}
\seealso{
\code{\link[=fmp_list_ciks]{fmp_list_ciks()}}
}
