% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmp_api_key}
\alias{fmp_api_key}
\title{Install a Financial Modeling Prep API Key in Your \code{.Renviron} File for repeated Use}
\usage{
fmp_api_key(key, install = TRUE, overwrite = FALSE)
}
\arguments{
\item{key}{\code{character}. The API key acquired from \href{https://financialmodelingprep.com/developer/docs/pricing/}{'https://financialmodelingprep.com/developer/docs/pricing/'}}

\item{install}{\code{logical}. If \code{TRUE}, will install the key in your \code{.Renviron} file for use in future sessions.
If \code{FALSE}, key is only loaded for the current R session and will not persist when a new session is loaded.
This is potentially useful if you wish to access the API from a non-personal computer. Default is \code{TRUE}.}

\item{overwrite}{\code{logical}. If \code{TRUE}, overwrite any existing FMP_API_KEY that you already have in your \code{.Renviron} file.}
}
\description{
This function will add your FMP API key to your \code{.Renviron} file
so it can be called securely without being stored in your code.
After you have installed your key, it can be called any time by
typing \code{Sys.getenv('FMP_API_KEY')}.
}
\examples{

\dontrun{
key <- "my_api_key"

fmp_api_key(key)

# first time, either reload your environment or restart R
# reload environment with the following
readRenviron("~/.Renviron")

# you can check it with:
Sys.getenv("FMP_API_KEY")
}
}
