% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-prices.R
\name{fmp_daily_prices}
\alias{fmp_daily_prices}
\title{Daily Security Prices}
\usage{
fmp_daily_prices(
  symbol,
  start_date = NULL,
  end_date = NULL,
  last_n = NULL,
  ...
)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}

\item{start_date}{\code{character}. ISO-8601 formatted date. For example, \code{'2020-05-25'}}

\item{end_date}{\code{character}. ISO-8601 formatted date. For example, \code{'2020-06-25'}}

\item{last_n}{\code{numeric}. An alternative to \code{start_date} and \code{end_date}. return the last \code{n} days}

\item{...}{additional parameters}
}
\value{
a \link[tibble:tibble-package]{tibble} of quotes of requested securities
}
\description{
Daily Security Prices
}
\examples{
\donttest{

# last 100 days of Apple stock prices
d <- fmp_daily_prices('AAPL', last_n = 100)
}

}
