% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-valuation.R
\name{fmp_dcf}
\alias{fmp_dcf}
\title{Company Discounted Cash Flow Value}
\usage{
fmp_dcf(symbol, historical = FALSE, quarterly = FALSE)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}

\item{historical}{\code{logical}. If \code{TRUE} return historical dcf values. If \code{FALSE}
return current estimate. Default is \code{FALSE}}

\item{quarterly}{\code{logical}. If \code{TRUE} return quarterly. If \code{FALSE}
return annual. Default is \code{FALSE}}
}
\value{
a \link[tibble:tibble-package]{tibble} of relevant financial information
}
\description{
Company Discounted Cash Flow Value
}
\examples{
my_stocks <- c('AAPL', 'GE')
d <- fmp_dcf(my_stocks, historical = TRUE, quarterly = TRUE)

}
\seealso{
Other \verb{Company Summaries}: 
\code{\link{fmp_balance_sheet}()},
\code{\link{fmp_cash_flow}()},
\code{\link{fmp_enterprise_value}()},
\code{\link{fmp_financial_growth}()},
\code{\link{fmp_income}()},
\code{\link{fmp_key_metrics}()},
\code{\link{fmp_profile}()},
\code{\link{fmp_quote}()},
\code{\link{fmp_rating}()},
\code{\link{fmp_ratios}()}
}
\concept{\verb{Company Summaries}}
