% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-valuation.R
\name{fmp_dividends}
\alias{fmp_dividends}
\title{Company Stock Dividends}
\usage{
fmp_dividends(symbol)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}
}
\value{
a \link[tibble:tibble-package]{tibble} of relevant financial information
}
\description{
Company Stock Dividends
}
\examples{

\donttest{
my_stocks <- c('AAPL', 'GE')
d <- fmp_dividends(my_stocks)
}

}
\seealso{
Other \verb{Company Summaries}: 
\code{\link{fmp_balance_sheet}()},
\code{\link{fmp_cash_flow}()},
\code{\link{fmp_dcf}()},
\code{\link{fmp_enterprise_value}()},
\code{\link{fmp_financial_growth}()},
\code{\link{fmp_full_financial}()},
\code{\link{fmp_income}()},
\code{\link{fmp_key_executives}()},
\code{\link{fmp_key_metrics}()},
\code{\link{fmp_market_cap}()},
\code{\link{fmp_profile}()},
\code{\link{fmp_rating}()},
\code{\link{fmp_ratios}()},
\code{\link{fmp_splits}()}
}
\concept{\verb{Company Summaries}}
