% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-valuation.R
\name{fmp_earnings}
\alias{fmp_earnings}
\title{Company Historical Earnings}
\usage{
fmp_earnings(symbol)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}
}
\value{
a \link[tibble:tibble-package]{tibble} of relevant financial information
}
\description{
Company Historical Earnings
}
\note{
Currently \code{fmp_earnings()} requires increased API access
}
