% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-valuation.R
\name{fmp_financial_growth}
\alias{fmp_financial_growth}
\title{Company Financial Growth}
\usage{
fmp_financial_growth(symbol, quarterly = FALSE)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}

\item{quarterly}{\code{logical}. If \code{TRUE} return quarterly. If \code{FALSE}
return annual. Default is \code{FALSE}}
}
\value{
a \link[tibble:tibble-package]{tibble} of relevant financial information
}
\description{
Company Financial Growth
}
\examples{

\donttest{
my_stocks <- c('AAPL', 'GE')
d <- fmp_financial_growth(my_stocks, quarterly = TRUE)
}

}
\seealso{
Other \verb{Company Summaries}: 
\code{\link{fmp_balance_sheet}()},
\code{\link{fmp_cash_flow}()},
\code{\link{fmp_company_outlook}()},
\code{\link{fmp_dcf}()},
\code{\link{fmp_dividends}()},
\code{\link{fmp_enterprise_value}()},
\code{\link{fmp_full_financial}()},
\code{\link{fmp_income}()},
\code{\link{fmp_institutional_holders}()},
\code{\link{fmp_key_executives}()},
\code{\link{fmp_key_metrics}()},
\code{\link{fmp_market_cap}()},
\code{\link{fmp_profile}()},
\code{\link{fmp_rating}()},
\code{\link{fmp_ratios}()},
\code{\link{fmp_shares_float}()},
\code{\link{fmp_splits}()}
}
\concept{\verb{Company Summaries}}
