% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-valuation.R
\name{fmp_income}
\alias{fmp_income}
\title{Company Income Statement}
\usage{
fmp_income(symbol, quarterly = FALSE)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}

\item{quarterly}{\code{logical}. If \code{TRUE} return quarterly. If \code{FALSE}
return annual. Default is \code{FALSE}}
}
\value{
a \link[tibble:tibble-package]{tibble} of relevant financial information
}
\description{
Company Income Statement
}
\examples{
my_stocks <- c('AAPL', 'GE')
d <- fmp_income(my_stocks, quarterly = TRUE)

}
\seealso{
Other \verb{Company Summaries}: 
\code{\link{fmp_balance_sheet}()},
\code{\link{fmp_cash_flow}()},
\code{\link{fmp_dcf}()},
\code{\link{fmp_enterprise_value}()},
\code{\link{fmp_financial_growth}()},
\code{\link{fmp_key_metrics}()},
\code{\link{fmp_profile}()},
\code{\link{fmp_quote}()},
\code{\link{fmp_rating}()},
\code{\link{fmp_ratios}()}
}
\concept{\verb{Company Summaries}}
