% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-list.R
\name{fmp_list}
\alias{fmp_list}
\alias{fmp_list_nyse}
\alias{fmp_list_nasdaq}
\alias{fmp_list_amex}
\alias{fmp_list_tsx}
\alias{fmp_list_etfs}
\alias{fmp_list_forex}
\alias{fmp_list_cryptos}
\alias{fmp_list_euronext}
\alias{fmp_list_mutual_funds}
\alias{fmp_list_commodities}
\alias{fmp_list_indexes}
\alias{fmp_list_ciks}
\alias{fmp_list_stocks}
\title{List Available Securities}
\usage{
fmp_list_nyse()

fmp_list_nasdaq()

fmp_list_amex()

fmp_list_tsx()

fmp_list_etfs()

fmp_list_forex()

fmp_list_cryptos()

fmp_list_euronext()

fmp_list_mutual_funds()

fmp_list_commodities()

fmp_list_indexes()

fmp_list_ciks()

fmp_list_stocks()
}
\value{
a \link[tibble:tibble-package]{tibble} of quotes of requested securities
}
\description{
Group of functions for listing available securities
}
\details{
Group of functions Details paragraph.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

\examples{

\donttest{
library(fmpapi)

# all available stocks
df_stocks <- fmp_list_stocks()

df_etfs <- fmp_list_etfs()
}
}
