% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-valuation.R
\name{fmp_market_cap}
\alias{fmp_market_cap}
\title{Company Market Capitalization}
\usage{
fmp_market_cap(symbol, historical = FALSE)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}

\item{historical}{\code{logical}. If \code{TRUE} return daily historical values. If \code{FALSE}
return current estimate. Default is \code{FALSE}}
}
\value{
a \link[tibble:tibble-package]{tibble} of relevant financial information
}
\description{
Company Market Capitalization
}
\examples{

\donttest{
my_stocks <- c('AAPL', 'GE')
d <- fmp_market_cap(my_stocks, historical = TRUE)
}

}
\seealso{
Other \verb{Company Summaries}: 
\code{\link{fmp_balance_sheet}()},
\code{\link{fmp_cash_flow}()},
\code{\link{fmp_dcf}()},
\code{\link{fmp_dividends}()},
\code{\link{fmp_enterprise_value}()},
\code{\link{fmp_financial_growth}()},
\code{\link{fmp_full_financial}()},
\code{\link{fmp_income}()},
\code{\link{fmp_key_executives}()},
\code{\link{fmp_key_metrics}()},
\code{\link{fmp_profile}()},
\code{\link{fmp_rating}()},
\code{\link{fmp_ratios}()},
\code{\link{fmp_splits}()}
}
\concept{\verb{Company Summaries}}
