% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-prices.R
\name{fmp_prices}
\alias{fmp_prices}
\title{Intra-day Security Prices}
\usage{
fmp_prices(
  symbol,
  interval = c("1min", "5min", "15min", "30min", "1hour", "4hour")
)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}

\item{interval}{\code{character}. Time resolution of price quotes}
}
\value{
a \link[tibble:tibble-package]{tibble} of quotes of requested securities
}
\description{
Intra-day Security Prices
}
\examples{
\donttest{

# hourly Apple stock prices
d <- fmp_prices('AAPL', interval = '1hour')
}

}
