% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-screener.R
\name{fmp_screen_stocks}
\alias{fmp_screen_stocks}
\title{Company Stock Screener}
\usage{
fmp_screen_stocks(
  marketCapMoreThan = NULL,
  marketCapLowerThan = NULL,
  betaMoreThan = NULL,
  betaLowerThan = NULL,
  volumeMoreThan = NULL,
  volumeLowerThan = NULL,
  dividendMoreThan = NULL,
  dividendLowerThan = NULL,
  sector = NULL,
  exchange = NULL,
  limit = NULL
)
}
\arguments{
\item{marketCapMoreThan}{\code{numeric}. Default is \code{NULL}}

\item{marketCapLowerThan}{\code{numeric}. Default is \code{NULL}}

\item{betaMoreThan}{\code{numeric}. Default is \code{NULL}}

\item{betaLowerThan}{\code{numeric}. Default is \code{NULL}}

\item{volumeMoreThan}{\code{numeric}. Default is \code{NULL}}

\item{volumeLowerThan}{\code{numeric}. Default is \code{NULL}}

\item{dividendMoreThan}{\code{numeric}. Default is \code{NULL}}

\item{dividendLowerThan}{\code{numeric}. Default is \code{NULL}}

\item{sector}{\code{character}. See Details for avialable values}

\item{exchange}{\code{character}. See Details for avialable values}

\item{limit}{\code{numeric}. Default is \code{NULL}}
}
\value{
a \link[tibble:tibble-package]{tibble} of stocks matching criteria
}
\description{
Company Stock Screener
}
\details{
\code{sector} paramater can be anyone of the following:
\itemize{
\item \code{"Consumer Cyclical"}
\item \code{"Energy"}
\item \code{"Technology"}
\item \code{"Industrials"}
\item \code{"Financial Services"}
\item \code{"Basic Materials"}
\item \code{"Communication Services"}
\item \code{"Consumer Defensive"}
\item \code{"Healthcare"}
\item \code{"Real Estate"}
\item \code{"Utilities"}
\item \code{"Industrial Goods"}
\item \code{"Financial"}
\item \code{"Services"}
\item \code{"Conglomerates"}
}

\code{exchange} paramater can be anyone of the following:
\itemize{
\item \code{"nyse"}
\item \code{"nasdaq"}
\item \code{"amex"}
\item \code{"euronext"}
\item \code{"tsx"}
\item \code{"etf"}
\item \code{"mutual_fund"}
}
}
\examples{

\donttest{
library(fmpapi)

# small cap, high beta tech stocks
d <- fmp_screen_stocks(marketCapLowerThan = 1e9, betaMoreThan = 2, sector = 'Technology')

# mid cap healthcare stocks listed on the nasdaq
d <- fmp_screen_stocks(marketCapMoreThan = 1e9, marketCapLowerThan = 1e10,
sector = 'Healthcare', exchange = 'nasdaq')
}

}
