% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutional-funds.R
\name{fmp_sec_filings}
\alias{fmp_sec_filings}
\title{SEC RSS Feed}
\usage{
fmp_sec_filings()
}
\value{
a \link[tibble:tibble-package]{tibble} of filing information
}
\description{
Retrieve the most recent financial statements filed to the SEC
}
