% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-market-info.R
\name{fmp_sector_performance}
\alias{fmp_sector_performance}
\title{Market Sector Performance}
\usage{
fmp_sector_performance(historical = FALSE)
}
\arguments{
\item{historical}{\code{logical}. If \code{TRUE} return historical values. If \code{FALSE}
return current. Default is \code{FALSE}}
}
\description{
Market Sector Performance
}
