% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmpapi-package.R
\docType{package}
\name{fmpapi-package}
\alias{fmpapi}
\alias{fmpapi-package}
\title{fmpapi: Access to the Financial Modeling Prep Financial Data API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Simple and consistent interface to the Financial Modeling Prep Financial Data API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpiburn/fmpapi}
  \item Report bugs at \url{https://github.com/jpiburn/fmpapi/issues}
}

}
\author{
\strong{Maintainer}: Jesse Piburn \email{jesse.piburn@gmail.com} (\href{https://orcid.org/0000-0003-4967-7912}{ORCID})

}
\keyword{internal}
