% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-quotes.R
\name{fmpQuote}
\alias{fmpQuote}
\alias{fmp_quote_nyse}
\alias{fmp_quote_nasdaq}
\alias{fmp_quote_amex}
\alias{fmp_quote_tsx}
\alias{fmp_quote_etfs}
\alias{fmp_quote_forex}
\alias{fmp_quote_cryptos}
\alias{fmp_quote_euronext}
\alias{fmp_quote_mutual_funds}
\alias{fmp_quote_commodities}
\alias{fmp_quote_indexes}
\title{Retrieve Real-Time Quotes of Various Securities and Cryptocurrencies}
\usage{
fmp_quote_nyse()

fmp_quote_nasdaq()

fmp_quote_amex()

fmp_quote_tsx()

fmp_quote_etfs()

fmp_quote_forex()

fmp_quote_cryptos()

fmp_quote_euronext()

fmp_quote_mutual_funds()

fmp_quote_commodities()

fmp_quote_indexes()
}
\value{
a \link[tibble:tibble-package]{tibble} of quotes of requested securities
}
\description{
Retrieve Real-Time Quotes of Various Securities and Cryptocurrencies
}
