% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutional-funds.R
\name{institutionalFund}
\alias{institutionalFund}
\alias{fmp_13f}
\title{Institutional Fund}
\usage{
fmp_13f(cik, year)
}
\arguments{
\item{cik}{\code{character}. A vector of Central Index Keys (CIK)}

\item{year}{\code{numeric}. 4 digit year, e.g. \code{2020}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

\examples{

berkshire_cik <- '0001067983'
d <- fmp_13f(berkshire_cik, year = 2020)
}
