% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-valuation.R
\name{companyValuation}
\alias{companyValuation}
\alias{fmp_profile}
\alias{fmp_quote}
\alias{fmp_income}
\alias{fmp_balance_sheet}
\alias{fmp_cash_flow}
\alias{fmp_ratios}
\alias{fmp_enterprise_value}
\alias{fmp_key_metrics}
\alias{fmp_financial_growth}
\alias{fmp_rating}
\alias{fmp_dcf}
\alias{fmp_earnings_calendar}
\alias{fmp_earnings}
\title{Company Valuation}
\usage{
fmp_profile(symbol)

fmp_quote(symbol)

fmp_income(symbol, quarterly = FALSE)

fmp_balance_sheet(symbol, quarterly = FALSE)

fmp_cash_flow(symbol, quarterly = FALSE)

fmp_ratios(symbol, quarterly = FALSE)

fmp_enterprise_value(symbol, quarterly = FALSE)

fmp_key_metrics(symbol, quarterly = FALSE)

fmp_financial_growth(symbol, quarterly = FALSE)

fmp_rating(symbol)

fmp_dcf(symbol, historical = FALSE, quarterly = FALSE)

fmp_earnings_calendar()

fmp_earnings(symbol)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}

\item{quarterly}{\code{logical}. If \code{TRUE} return quarterly. If \code{FALSE}
return annual. Default is \code{FALSE}}

\item{historical}{\code{logical}. If \code{TRUE} return historical dcf values. If \code{FALSE}
return current estimate. Default is \code{FALSE}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\note{
Currently \code{fmp_earnings()} requirest increased API access
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

\examples{
my_stocks <- c('AAPL', 'GE')

d <- fmp_profile(my_stocks)
d <- fmp_quote(my_stocks)
d <- fmp_income(my_stocks)
d <- fmp_balance_sheet(my_stocks)
d <- fmp_cash_flow(my_stocks)
d <- fmp_ratios(my_stocks)
d <- fmp_enterprise_value(my_stocks)
d <- fmp_key_metrics(my_stocks)
d <- fmp_financial_growth(my_stocks)
d <- fmp_rating(my_stocks)
d <- fmp_dcf(my_stocks, historical = TRUE, quarterly = TRUE)
d <- fmp_earnings_calendar()
}
