% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-package.R
\docType{package}
\name{fmp-package}
\alias{fmp}
\alias{fmp-package}
\title{fmp: Access to the Financial Modeling Prep Financial Data API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Simple and consistent interface to the Financial Modeling Prep Financial Data API.
}
\author{
\strong{Maintainer}: Jesse Piburn \email{piburnjo@ornl.gov} (\href{https://orcid.org/0000-0003-4967-7912}{ORCID})

Other contributors:
\itemize{
  \item UT-Battelle, LLC [copyright holder]
}

}
\keyword{internal}
