% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp-quotes.R
\name{fmp_quote}
\alias{fmp_quote}
\title{Retrieve Real-Time Quotes of Selected Securities and Cryptocurrencies}
\usage{
fmp_quote(symbol)
}
\arguments{
\item{symbol}{\code{character}. A vector of stock symbols.}
}
\value{
a \link[tibble:tibble-package]{tibble} of quotes of requested securities
}
\description{
Retrieve Real-Time Quotes of Selected Securities and Cryptocurrencies
}
\examples{

\donttest{
my_stocks <- c('AAPL', 'GE')
d <- fmp_quote(my_stocks)
}

}
